выделлил общие части компонентов в абстрактные классы
