--- src/util.js
@@ -1,30 +1,0 @@
-import {RenderPosition} from './constants.js';
-
-const getRandomBetween = (min, max) => {
-  return Math.floor(Math.random() * (max - min) + min);
-};
-
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomBetween(0, array.length);
-
-  return array[randomIndex];
-};
-
-const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
-
-const renderHtmlPart = (container, template, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(template);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(template);
-      break;
-  }
-};
-
-export {getRandomBetween, getRandomArrayItem, createElement, renderHtmlPart};
